requirements.txt
akshare==1.1.18
alabaster==0.7.12
allure-pytest==2.9.43
allure-python-commons==2.9.43
Appium-Python-Client==1.2.0
argcomplete==1.12.3
argon2-cffi==21.1.0
atomicwrites==1.4.0
attrs==21.2.0
Automat==20.2.0
Babel==2.9.1
backcall==0.2.0
beautifulsoup4==4.9.3
bleach==4.1.0
bs4==0.0.1
cachetools==4.2.2
certifi==2021.5.30
cffi==1.14.6
charset-normalizer==2.0.4
click==8.0.1
colorama==0.4.4
comtypes==1.1.10
constantly==15.1.0
cryptography==3.4.8
cssselect==1.1.0
cssutils==2.3.0
cycler==0.10.0
ddddocr==1.0.8
ddt==1.4.2
debugpy==1.4.3
decorator==5.1.0
defusedxml==0.7.1
demjson==2.2.4
docutils==0.17.1
entrypoints==0.3
et-xmlfile==1.1.0
Faker==8.12.1
flatbuffers==2.0
ghp-import==2.0.1
h2==3.2.0
hpack==3.0.0
html5lib==1.1
hyperframe==5.2.0
hyperlink==21.0.0
idna==3.2
imagesize==1.2.0
importlib-metadata==4.8.1
incremental==21.3.0
iniconfig==1.1.1
ipykernel==6.4.1
ipython==7.27.0
ipython-genutils==0.2.0
ipywidgets==7.6.5
itemadapter==0.4.0
itemloaders==1.0.4
jedi==0.18.0
Jinja2==3.0.1
jmespath==0.10.0
jsonpath==0.82
jsonschema==3.2.0
jupyter==1.0.0
jupyter-client==7.0.3
jupyter-console==6.4.0
jupyter-core==4.8.1
jupyterlab-pygments==0.1.2
jupyterlab-widgets==1.0.2
kiwisolver==1.3.2
loguru==0.5.3
lxml==4.6.3
Markdown==3.3.4
MarkupSafe==2.0.1
matplotlib==3.4.3
matplotlib-inline==0.1.3
mergedeep==1.3.4
mistune==0.8.4
mkdocs==1.2.2
MouseInfo==0.1.3
nbclient==0.5.4
nbconvert==6.1.0
nbformat==5.1.3
nest-asyncio==1.5.1
notebook==6.4.4
numpy==1.21.2
onnxruntime==1.8.1
openpyxl==3.0.8
packaging==21.0
pandas==1.3.2
pandocfilters==1.5.0
parseIdCard==1.0.0
parsel==1.6.0
parso==0.8.2
pickleshare==0.7.5
Pillow==8.3.1
pluggy==1.0.0
premailer==3.10.0
priority==1.3.0
prometheus-client==0.11.0
prompt-toolkit==3.0.20
Protego==0.1.16
protobuf==3.17.3
py==1.10.0
py-mini-racer==0.6.0
pyasn1==0.4.8
pyasn1-modules==0.2.8
PyAutoGUI==0.9.53
pycparser==2.20
PyDispatcher==2.0.5
PyGetWindow==0.0.9
Pygments==2.10.0
PyMsgBox==1.0.9
pyOpenSSL==20.0.1
pyparsing==2.4.7
pyperclip==1.8.2
pypinyin==0.42.0
pyquery==1.4.3
PyRect==0.1.4
pyrsistent==0.18.0
PyScreeze==0.1.27
PySnooper==0.5.0
pytest==6.2.5
pytest-assume==2.4.3
pytest-dependency==0.5.1
pytest-ordering==0.6
python-dateutil==2.8.2
PyTweening==1.0.3
pytz==2021.1
pywin32==301
pywinauto==0.6.8
pywinpty==1.1.4
PyYAML==5.4.1
pyyaml_env_tag==0.1
pyzmq==22.3.0
qtconsole==5.1.1
QtPy==1.11.1
queuelib==1.6.2
requests==2.26.0
Scrapy==2.5.0
selenium==3.141.0
Send2Trash==1.8.0
service-identity==21.1.0
six==1.16.0
snowballstemmer==2.1.0
soupsieve==2.2.1
Sphinx==4.1.2
sphinxcontrib-applehelp==1.0.2
sphinxcontrib-devhelp==1.0.2
sphinxcontrib-htmlhelp==2.0.0
sphinxcontrib-jsmath==1.0.1
sphinxcontrib-qthelp==1.0.3
sphinxcontrib-serializinghtml==1.1.5
tabulate==0.8.9
terminado==0.12.1
testpath==0.5.0
text-unidecode==1.3
toml==0.10.2
tornado==6.1
tqdm==4.62.2
traitlets==5.1.0
Twisted==21.7.0
twisted-iocpsupport==1.0.2
typing-extensions==3.10.0.2
urllib3==1.25.8
vboxapi==1.0
w3lib==1.22.0
watchdog==2.1.5
wcwidth==0.2.5
webencodings==0.5.1
websocket-client==1.2.1
widgetsnbextension==3.5.1
win32-setctime==1.0.3
xlrd==2.0.1
xlwings==0.24.9
yagmail==0.14.260
you-get==0.4.1545
zipp==3.5.0
zope.interface==5.4.0
